Immediate logfile creation
This commit creates the log files immediately in the parent process
to guarantee that that the same jid is not claimed but several
child processes. Before, if the creation of the logger was delayed
in the callable, the same jid could be given to several child process,
which would overwrite each other.